+++ beacon-chain/state/state-native/v2/getters_validator_test.go
@@ -4,18 +4,15 @@ import (
 	"testing"
 
 	"github.com/prysmaticlabs/prysm/beacon-chain/state"
-	v1 "github.com/prysmaticlabs/prysm/beacon-chain/state/state-native/v1"
+	testtmpl "github.com/prysmaticlabs/prysm/beacon-chain/state/testing"
+	v2 "github.com/prysmaticlabs/prysm/beacon-chain/state/v2"
 	ethpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1"
-	"github.com/prysmaticlabs/prysm/testing/assert"
-	"github.com/prysmaticlabs/prysm/testing/require"
 )
 
 func TestBeaconState_ValidatorAtIndexReadOnly_HandlesNilSlice(t *testing.T) {
-	st, err := v1.InitializeFromProtoUnsafe(&ethpb.BeaconState{
-		Validators: nil,
+	testtmpl.VerifyBeaconState_ValidatorAtIndexReadOnly_HandlesNilSlice(t, func() (state.BeaconState, error) {
+		return v2.InitializeFromProtoUnsafe(&ethpb.BeaconStateAltair{
+			Validators: nil,
+		})
 	})
-	require.NoError(t, err)
-
-	_, err = st.ValidatorAtIndexReadOnly(0)
-	assert.Equal(t, state.ErrNilValidatorsInState, err)
 }
