+++ beacon-chain/state/state-native/v3/BUILD.bazel
@@ -85,8 +85,8 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/state:go_default_library",
-        "//beacon-chain/state/state-native/v1:go_default_library",
         "//beacon-chain/state/stateutil:go_default_library",
+        "//beacon-chain/state/testing:go_default_library",
         "//beacon-chain/state/v3:go_default_library",
         "//config/features:go_default_library",
         "//config/fieldparams:go_default_library",
