+++ beacon-chain/state/testing/getters_validator.go
@@ -1,0 +1,19 @@
+package testing
+
+import (
+	"testing"
+
+	"github.com/prysmaticlabs/prysm/beacon-chain/state"
+	"github.com/prysmaticlabs/prysm/testing/assert"
+	"github.com/prysmaticlabs/prysm/testing/require"
+)
+
+type getState func() (state.BeaconState, error)
+
+func VerifyBeaconState_ValidatorAtIndexReadOnly_HandlesNilSlice(t *testing.T, factory getState) {
+	st, err := factory()
+	require.NoError(t, err)
+
+	_, err = st.ValidatorAtIndexReadOnly(0)
+	assert.Equal(t, state.ErrNilValidatorsInState, err)
+}
